refactor(config): extract protocol scalar rewriting helpers
Problem: The expanded protocol migration exceeded the repository file-size gate.

Root cause: TOML scalar rewriting and provider migration shared one file near its size ceiling.

Fix: Move the existing scalar rewriting functions into a dedicated file without changing behavior or increasing the lint baseline.

Verification: repolint is clean and the configuration suite passes. Root, desktop, frontend build and migration race checks passed before this mechanical extraction.
